Quake 4 BSE effects update 7: experimental: sound broken
Because sound emitter will be free on rvClientEffect destroy on game, and it also destroy on BSE effect when rvClientEffect destroy. So it will call FreeSoundEmitter twice on one handle.
When call rvClientEffect destructor, it will be synchronous and safety.
But BSE effect free immediate first and then rvClientEffect will be removed delayed when call UpdateEffectDef return true in rvClientEffect::Think.
I think it maybe happen that:
Frame 1: BSE free a sound emitter handle -> rvClientEffect removed delayed(will be removed next frame, but not removed in this frame) -> other idEntity alloc this sount emitter handle
Frame 2: rvClientEffect id free, so this sound emitter handle is free -> but other idEntity is using this sound emitter handle